Absorb DatabaseMigrationStepApplier into MigrationConnector
- Using concrete trait objects for namespacing is an antipattern
- This is extra noise
- The trait has reduced enough that we can take the three extra
  top-level methods.
- Documentation on the trait was plain wrong.

Note that in the spirit of simplicity, we also return explicit futures
instead of async trait. Let's move away from async-trait in
migration-connector.
