+++ migration-engine/core/src/commands/diff.rs
@@ -30,10 +30,7 @@ pub(crate) async fn diff(params: DiffParams, host: Arc<dyn ConnectorHost>) -> Co
     let migration: migration_connector::Migration = connector.diff(from.target, to.target).await?;
 
     if params.script {
-        let script_string = connector.database_migration_step_applier().render_script(
-            &migration,
-            &migration_connector::DestructiveChangeDiagnostics::default(),
-        )?;
+        let script_string = connector.render_script(&migration, &Default::default())?;
         connector.host().print(&script_string).await?;
     } else {
         let summary = connector.migration_summary(&migration);
